backend/requirements.txt
Flask==3.0.3
flask-cors==4.0.1
SQLAlchemy==2.0.31
flask_sqlalchemy==3.1.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
stripe==9.9.0
firebase-admin==6.5.0
